dev-environment-templates/examples/python-example/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
